muggle/requirements.txt
httpcore==0.15
scikit-build
certifi
onnxruntime
numpy
Pillow==8.4.0
opencv-python-headless
pyyaml
requests
tinyaes
certifi
requests
werkzeug
paramiko
six
nuitka
zstandard
pyyaml
easycython
cryptography
requests
loguru
pyOpenSSL
pycryptodome
ordered-set
waitress
easycython
cffi
gradio
markdown
fastapi
matplotlib
uvicorn
pytz
jinja2
psutil